chore(types): re-order types per module
